feat(components): adress review
